// apps/trade-pro/src/pages/Taxation/AddTaxShedule/type.ts
export type TAddTaxSheduleHistory = {

  Id: number;
  TaxNameId: number;
  EffectedDate: Date;
  EntryDate: Date;
  EntryUser: number;
  ModifyDate: Date;
  ModifyUser: number;
  PostDate: Date;
  PostUser: number;
  PostState: Boolean;
  OrganizationId: number;
  CompanyId: number;
  TaxPercent: number;
  TaxName: string;
};